+++ src/routes/Font_Weight_Control.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	// TODO @multiple publish in $lib when ready
 
-	export let selected_font_weight = 500;
+	export let selected_font_weight = 400;
 </script>
 
 <label